feat: move error and label
